pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "guildgreeter"
version = "1.0.0"
description = "Bot Discord de bienvenue avec dashboard"
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3.2",
    "py-cord>=2.4.1",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "SQLAlchemy>=2.0.23",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "pyjwt>=2.8.0",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.1",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "psutil>=5.9.7",
    "aioredis>=2.0.1",
]

[tool.setuptools.packages.find]
where = ["src"]